Fixed OpenMP usage on diverse platforms, some uint changed to size_t
git-svn-id: https://svn.vital-it.ch/svn/hp2c/trunk/Codeml_Variants/Fastcodeml@1788 95c5a404-1f88-0410-a6b0-c3f062b6f34a

+++ Genes.h
@@ -38,7 +38,7 @@ public:
 	///
 	/// @return The number of sites
 	///
-	unsigned int getNumSites(void) const {return mSiteMultiplicity.size();}
+	size_t getNumSites(void) const {return mSiteMultiplicity.size();}
 
 	/// Return the site multiplicity
 	///
